Add "end of" embedding lookup support to LiteRT LM
There are special end of tokens that need to be looked up at the end
of multi-modal streams. For example, in some cases at the end of a
series of audio tokens, we need to append a special "end of audio"
token. This is looked up in a tiny embedding table.

LiteRT-LM-PiperOrigin-RevId: 792707891