Add synchronization for WebGPU Accelerator
Invoke WaitForCompletion() when all output tensors are external tensors
in NoExternalTensorsMode.
Using outputs of all external tensors with NoExternalTensorsMode looks contradict.
This can be happen when outputs are all state tensors and they'll be used as
inputs of the next execution with a different subgraph.

Use the WaitForCompletion() until we found better synchronization method for
WebGPU.

LiteRT-LM-PiperOrigin-RevId: 802630958